Move Coco code around
