+++ typescript/web-app/src/connectors/resolvers/format-coco/__tests__/export-to-coco.ts
@@ -1,10 +1,10 @@
 import gql from "graphql-tag";
-import { client } from "../../apollo-client-schema";
+import { client } from "../../../apollo-client-schema";
 
-import { setupTestsWithLocalDatabase } from "../../../utils/setup-local-db-tests";
-import { initialCocoDataset } from "../../../data-converters/coco-format/converters";
-import { CocoDataset } from "../../../data-converters/coco-format/types";
-import { jsonToDataUri } from "../export-to-coco";
+import { setupTestsWithLocalDatabase } from "../../../../utils/setup-local-db-tests";
+import { initialCocoDataset } from "../coco-core/converters";
+import { CocoDataset } from "../coco-core/types";
+import { jsonToDataUri } from "..";
 
 setupTestsWithLocalDatabase();
