+++ typescript/web-app/src/connectors/resolvers/format-coco/coco-core/types.ts
