+++ typescript/web-app/src/connectors/resolvers/format-coco/index.ts
@@ -1,0 +1,9 @@
+import { exportToCoco } from "./export-to-coco";
+
+export { jsonToDataUri } from "./json-to-data-uri";
+
+export default {
+  Query: {
+    exportToCoco,
+  },
+};
