+++ typescript/web-app/src/connectors/resolvers/format-coco/json-to-data-uri.ts
@@ -1,0 +1,2 @@
+export const jsonToDataUri = (json: string): string =>
+  `data:application/json;base64,${btoa(json)}`;
