+++ typescript/web-app/src/connectors/resolvers/index.ts
@@ -1,6 +1,6 @@
 import { mergeResolvers } from "@graphql-tools/merge";
 import exampleResolvers from "./example";
-import exportToCocoResolvers from "./export-to-coco";
+import formatCocoResolvers from "./format-coco";
 import imageResolvers from "./image";
 import labelResolvers from "./label";
 import labelClassResolvers from "./label-class";
@@ -9,7 +9,7 @@ import uploadResolvers from "./upload";
 
 export const resolvers = mergeResolvers([
   exampleResolvers,
-  exportToCocoResolvers,
+  formatCocoResolvers,
   imageResolvers,
   labelResolvers,
   labelClassResolvers,
