--- typescript/web-app/src/data-converters/README.md
@@ -1,4 +1,0 @@
-# Data converters
-
-This package provide a set of useful function to convert Labelflow entities to
-other formats.
